GHC.TypeLits.KnownNat

Description

Some "magic" classes and instances to get the GHC.TypeLits.KnownNat.Solver type checker plugin working.

Usage

Let's say you defined a closed type family Max:

import Data.Type.Bool (If)
import GHC.TypeLits

type family Max (a :: Nat) (b :: Nat) :: Nat where
  Max 0 b = b
  Max a b = If (a <=? b) b a

if you then want the GHC.TypeLits.KnownNat.Solver to solve KnownNat constraints over Max, given just KnownNat constraints for the arguments of Max, then you must define:

{-# LANGUAGE DataKinds, FlexibleInstances, GADTs, KindSignatures,
             MultiParamTypeClasses, ScopedTypeVariables, TemplateHaskell,
             TypeApplications, TypeFamilies, TypeOperators,
             UndecidableInstances #-}

import Data.Proxy            (Proxy (..))
import GHC.TypeLits.KnownNat

instance (KnownNat a, KnownNat b) => KnownNat2 $(nameToSymbol ''Max) a b where
  natSing2 = let x = natVal (Proxy a)
                 y = natVal (Proxy b)
                 z = max x y
             in  SNatKn z
  {-# INLINE natSing2 #-}

FAQ

1. GHC.TypeLits.KnownNat.Solver does not seem to find the corresponding KnownNat2 instance for my type-level operation

At the Core-level, GHCs internal mini-Haskell, type families that only have a single equation are treated like type synonyms.

For example, let's say we defined a closed type family Max:

import Data.Type.Bool (If)
import GHC.TypeLits

type family Max (a :: Nat) (b :: Nat) :: Nat where
  Max a b = If (a <=? b) b a

Now, a Haskell-level program might contain a constraint

KnownNat (Max a b)

, however, at the Core-level, this constraint is expanded to:

KnownNat (If (a <=? b) b a)

GHC.TypeLits.KnownNat.Solver never sees any reference to the Max type family, so it will not look for the corresponding KnownNat2 instance either. To fix this, ensure that your type-level operations always have at least two equations. For Max this means we have to redefine it as:

type family Max (a :: Nat) (b :: Nat) :: Nat where
  Max 0 b = b
  Max a b = If (a <=? b) b a

Singleton natural number

newtype SNatKn (f :: Symbol)

Singleton natural number

Constructors

SNatKn Natural

Constraint-level arithmetic classes

class KnownNat1 (f :: Symbol) (a :: Nat) where

Class for arithmetic functions with one argument.

The Symbol f must correspond to the fully qualified name of the type-level operation. Use nameToSymbol to get the fully qualified TH Name as a Symbol

Methods

natSing1 :: SNatKn f

class KnownNat2 (f :: Symbol) (a :: Nat) (b :: Nat) where

Class for arithmetic functions with two arguments.

The Symbol f must correspond to the fully qualified name of the type-level operation. Use nameToSymbol to get the fully qualified TH Name as a Symbol

Methods

natSing2 :: SNatKn f

class KnownNat3 (f :: Symbol) (a :: Nat) (b :: Nat) (c :: Nat) where

Class for arithmetic functions with three arguments.

The Symbol f must correspond to the fully qualified name of the type-level operation. Use nameToSymbol to get the fully qualified TH Name as a Symbol

Methods

natSing3 :: SNatKn f

Singleton boolean

data SBool (b :: Bool) where

Singleton version of Bool

Constructors

SFalse :: SBool 'False

STrue :: SBool 'True

boolVal :: forall b proxy. KnownBool b => proxy b -> Bool

Get the Bool value associated with a type-level Bool

Use boolVal if you want to perform the standard boolean operations on the reified type-level Bool.

Use boolSing if you need a context in which the type-checker needs the type-level Bool to be either True or False

f :: forall proxy b r . KnownBool b => r
f = case boolSing @b of
  SFalse -> -- context with b ~ False
  STrue  -> -- context with b ~ True

KnownBool

class KnownBool (b :: Bool) where

Methods

boolSing :: SBool b

Constraint-level boolean functions

newtype SBoolKb (f :: Symbol)

A type "representationally equal" to SBool, used for simpler implementation of constraint-level functions that need to create instances of KnownBool

Constructors

SBoolKb Bool

class KnownNat2Bool (f :: Symbol) (a :: Bool) (b :: k) (c :: k) where

Class for ternary functions with a Natural result.

The Symbol f must correspond to the fully qualified name of the type-level operation. Use nameToSymbol to get the fully qualified TH Name as a Symbol

Methods

natBoolSing3 :: SNatKn f

class KnownBoolNat2 (f :: Symbol) (a :: k) (b :: k) where

Class for binary functions with a Boolean result.

The Symbol f must correspond to the fully qualified name of the type-level operation. Use nameToSymbol to get the fully qualified TH Name as a Symbol

Methods

boolNatSing2 :: SBoolKb f

Template Haskell helper

nameToSymbol :: Name -> TypeQ

Convert a TH Name to a type-level Symbol